+++ src/binaryen-c.cpp
@@ -5373,6 +5373,18 @@ bool BinaryenGetDebugInfo(void) { return globalPassOptions.debugInfo; }
 
 void BinaryenSetDebugInfo(bool on) { globalPassOptions.debugInfo = on != 0; }
 
+bool BinaryenGetTrapsNeverHappen(void) {
+  return globalPassOptions.trapsNeverHappen;
+}
+
+void BinaryenSetTrapsNeverHappen(bool on) {
+  globalPassOptions.trapsNeverHappen = on;
+}
+
+bool BinaryenGetClosedWorld(void) { return globalPassOptions.closedWorld; }
+
+void BinaryenSetClosedWorld(bool on) { globalPassOptions.closedWorld = on; }
+
 bool BinaryenGetLowMemoryUnused(void) {
   return globalPassOptions.lowMemoryUnused;
 }
@@ -5392,6 +5404,22 @@ void BinaryenSetZeroFilledMemory(bool on) {
 bool BinaryenGetFastMath(void) { return globalPassOptions.fastMath; }
 
 void BinaryenSetFastMath(bool value) { globalPassOptions.fastMath = value; }
+
+bool BinaryenGetGenerateStackIR(void) {
+  return globalPassOptions.generateStackIR;
+}
+
+void BinaryenSetGenerateStackIR(bool on) {
+  globalPassOptions.generateStackIR = on;
+}
+
+bool BinaryenGetOptimizeStackIR(void) {
+  return globalPassOptions.optimizeStackIR;
+}
+
+void BinaryenSetOptimizeStackIR(bool on) {
+  globalPassOptions.optimizeStackIR = on;
+}
 
 const char* BinaryenGetPassArgument(const char* key) {
   assert(key);
@@ -5414,6 +5442,18 @@ void BinaryenSetPassArgument(const char* key, const char* value) {
 }
 
 void BinaryenClearPassArguments(void) { globalPassOptions.arguments.clear(); }
+
+bool BinaryenHasPassToSkip(const char* pass) {
+  assert(pass);
+  return globalPassOptions.passesToSkip.count(pass);
+}
+
+void BinaryenAddPassToSkip(const char* pass) {
+  assert(pass);
+  globalPassOptions.passesToSkip.insert(pass);
+}
+
+void BinaryenClearPassesToSkip(void) { globalPassOptions.passesToSkip.clear(); }
 
 BinaryenIndex BinaryenGetAlwaysInlineMaxSize(void) {
   return globalPassOptions.inlining.alwaysInlineMaxSize;
@@ -5453,7 +5493,10 @@ void BinaryenModuleRunPasses(BinaryenModuleRef module,
   PassRunner passRunner((Module*)module);
   passRunner.options = globalPassOptions;
   for (BinaryenIndex i = 0; i < numPasses; i++) {
-    passRunner.add(passes[i]);
+    passRunner.add(passes[i],
+                   globalPassOptions.arguments.count(passes[i]) > 0
+                     ? globalPassOptions.arguments[passes[i]]
+                     : std::optional<std::string>());
   }
   passRunner.run();
 }
@@ -5704,7 +5747,10 @@ void BinaryenFunctionRunPasses(BinaryenFunctionRef func,
   PassRunner passRunner((Module*)module);
   passRunner.options = globalPassOptions;
   for (BinaryenIndex i = 0; i < numPasses; i++) {
-    passRunner.add(passes[i]);
+    passRunner.add(passes[i],
+                   globalPassOptions.arguments.count(passes[i]) > 0
+                     ? globalPassOptions.arguments[passes[i]]
+                     : std::optional<std::string>());
   }
   passRunner.runOnFunction((Function*)func);
 }