+++ src/passes/LegalizeJSInterface.cpp
@@ -64,9 +64,8 @@ struct LegalizeJSInterface : public Pass {
     setTempRet0 = nullptr;
     getTempRet0 = nullptr;
     auto exportOriginals =
-      getPassOptions().hasArgument("legalize-js-interface-export-originals");
-    exportedHelpers =
-      getPassOptions().hasArgument("legalize-js-interface-exported-helpers");
+      hasArgument("legalize-js-interface-export-originals");
+    exportedHelpers = hasArgument("legalize-js-interface-exported-helpers");
     // for each illegal export, we must export a legalized stub instead
     std::vector<std::unique_ptr<Export>> newExports;
     for (auto& ex : module->exports) {