+++ test/binaryen.js/optimize-stack-ir.js
@@ -1,0 +1,3 @@
+console.log("// optimizeStackIR=" + binaryen.getOptimizeStackIR());
+binaryen.setOptimizeStackIR(true);
+assert(binaryen.getOptimizeStackIR() == true);